Add reading from Realm in FriendsService
